ENH Uses folder structure of packaging
